Add some basic plugins to vim configuration
Add some plugins to vim configuration for working on the server:
vim-fugitive and vim-gitgutter for git signs and git management, and
vim-commentary to easily comment or uncomment lines.
Plugins are installed using Plug. Plug is automatically downloaded using
wget (instead of curl, because wget is usually preinstalled in most
distros).